flask_app/bots.py
class PokerBot:
    def __init__(self, bot_id, name):
        self.bot_id = bot_id
        self.name = name
        self.hand = []
        self.chips = 1000
        self.bluff_factor = 0.1
        self.learning_rate = 0.01
        self.strategy = {} 
        self.history = []

    def decide_action(self, game_state):
        import random
        hand_strength = evaluate_hand(self.hand)
        state_key = self.get_state_key(game_state, hand_strength)
        if state_key in self.strategy:
            action = self.strategy[state_key]
        else:
            if random.random() < self.bluff_factor:
                action = 'raise'
            else:
                action = 'call' if hand_strength > 0.5 else 'fold'
            self.strategy[state_key] = action
        return {'action': action, 'amount': min(self.chips, 20) if action == 'raise' else 0}

    def get_state_key(self, game_state, hand_strength):
        return f"{hand_strength}-{game_state['pot_size']}"

    def receive_hand(self, hand):
        self.hand = hand

    def adjust_chips(self, amount):
        self.chips += amount

    def learn(self, outcome):
        reward = 1 if outcome == 'win' else -1
        for state_key, action in self.history:
            if state_key in self.strategy:
                # Update the strategy based on the reward
                current_value = self.strategy[state_key]
                updated_value = current_value + self.learning_rate * (reward - current_value)
                self.strategy[state_key] = updated_value
        
        self.history.clear() 

def bot_actions():
    for bot in bots:
        decision = bot.decide_action(game_state={})
        action = decision['action']
        if action == 'raise':
            amount = decision['amount']
            bot.adjust_chips(-amount)
            pot_size += amount
        elif action == 'call':
            bot.adjust_chips(-5)
            pot_size += 5
        # Other actions: check, fold
